Add test for --fit
