+++ tests/data/maps/world-test-2.xml.tmpl
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE Map[]>
+<Map srs="{{ web_mercator_srs }}" minimum-version="0.7.2" buffer-size="128" maximum-extent="-20037508.34,-20037508.34,20037508.34,20037508.34">
+  <FontSet name="fontset-0">
+    <Font face-name="Noto Sans Regular" />
+    <Font face-name="DejaVu Sans Book" />
+    <Font face-name="HanaMinA Regular" />
+    <Font face-name="HanaMinB Regular" />
+    <Font face-name="Unifont Medium" />
+    <Font face-name="unifont Medium" />
+    <Font face-name="Unifont Upper Medium" />
+  </FontSet>
+  <Style name="style">
+     <Rule>
+        <PolygonSymbolizer fill="yellow"/>
+     </Rule>
+   </Style>
+
+  <Layer name="countries" srs="{{ web_mercator_srs }}">
+    <StyleName>style</StyleName>
+    <Datasource>
+      <Parameter name="type">shape</Parameter>
+      <Parameter name="file">../shp/world_merc.shp</Parameter>
+    </Datasource>
+  </Layer>
+  <Style name="cities">
+    <Rule>
+       <TextSymbolizer placement="point" size="11" fill="black" fontset-name="fontset-0" halo-fill="rgba(255, 255, 255, 0.6)" halo-radius="1.5"><![CDATA[[name]]]></TextSymbolizer>
+    </Rule>
+  </Style>
+  <Layer name="cities" srs="{{ web_mercator_srs }}">
+    <StyleName>cities</StyleName>
+    <Datasource>
+      <Parameter name="type">shape</Parameter>
+      <Parameter name="file">../shp/labels.shp</Parameter>
+    </Datasource>
+  </Layer>
+</Map>
+
