+++ tests/data/shp/labels.cpg
@@ -1,0 +1,1 @@
+UTF-8
